requirements.txt
# Core RAG dependencies
langchain==0.1.0
langchain-community==0.0.13
langchain-openai==0.0.5
sentence-transformers==2.2.2
faiss-cpu==1.7.4
transformers==4.36.0

# PDF processing
PyPDF2==3.0.1
PyMuPDF
pdfplumber==0.10.0

# Web framework
fastapi==0.104.0
uvicorn==0.24.0
pydantic==2.5.0

# Bengali language support
bengali-stemmer==1.0.0
indic-nlp-library==0.81

# Database
chromadb==0.4.0
sqlite3

# General utilities
python-dotenv==1.0.0
numpy==1.24.0
pandas==2.1.0
requests==2.31.0
python-multipart==0.0.6

# Evaluation metrics
scikit-learn==1.3.0
rouge-score==0.1.2
bert-score==0.3.13

# Development
pytest==7.4.0
black==23.0.0
flake8==6.0.0
